+++ src/gamelogic/WorldParser.hpp
@@ -23,8 +23,8 @@ namespace GameLogic {
     private:
         RapidXMLAdapter _rapid_adapter;
         vector<vector<Tile*>> _generate_tilemap( vector<vector<int>> int_map, Engine::SPTR_AssetsManager assets );
-        CoordinateDouble _get_spawnpoint( vector<tuple<int, int, char*>> object_list, vector<vector<Tile*>> map );
-        void _set_objects( Level& level, vector<tuple<int, int, char*>> object_list, vector<vector<Tile*>> map, Engine::SPTR_AssetsManager assets, string file_location );
+        CoordinateDouble _get_spawnpoint( vector<tuple<size_t, size_t, char*>> object_list, vector<vector<Tile*>> map );
+        void _set_objects( Level& level, vector<tuple<size_t, size_t, char*>> object_list, vector<vector<Tile*>> map, Engine::SPTR_AssetsManager assets, string file_location );
         double _spawn_tile_offset = 0.5;
         string _path;
     };
